+++ pkg/agent/agent_mcp.go
